# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/subscribers/{id}:
    get:
      tags:
      - Management-data-controller
      operationId: getSubscriber
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubscriberResponse"
    put:
      tags:
      - Management-data-controller
      operationId: updateSubscriber
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubscriberResponse"
    delete:
      tags:
      - Management-data-controller
      operationId: deleteSubscriber
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/subscriber-features/{id}:
    get:
      tags:
      - Management-data-controller
      operationId: getSubscriberFeature
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubscriberFeatureResponse"
    put:
      tags:
      - Management-data-controller
      operationId: updateSubscriberFeature
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberFeatureUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubscriberFeatureResponse"
    delete:
      tags:
      - Management-data-controller
      operationId: deleteSubscriberFeature
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/additional-features/{id}:
    get:
      tags:
      - Management-data-controller
      operationId: getService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdditionalFeatureResponse"
    put:
      tags:
      - Management-data-controller
      operationId: updateService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdditionalFeatureUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdditionalFeatureResponse"
    delete:
      tags:
      - Management-data-controller
      operationId: deleteService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/feature-categories/{id}:
    get:
      tags:
      - Management-data-controller
      operationId: getServiceCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdNameResponse"
    put:
      tags:
      - Management-data-controller
      operationId: updateServiceCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdNameResponse"
    delete:
      tags:
      - Management-data-controller
      operationId: deleteServiceCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/v1/subscribers/{subscriberId}/features/{featureId}/disable:
    post:
      tags:
      - service-controller
      operationId: disableFeature
      parameters:
      - name: subscriberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: featureId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DisableFeatureResponse"
...
  /api/v1/subscribers:
    get:
      tags:
      - Management-data-controller
      operationId: getSubscribers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriberResponse"
    post:
      tags:
      - Management-data-controller
      operationId: createSubscriber
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubscriberResponse"
  /api/v1/subscriber-features:
    get:
      tags:
      - Management-data-controller
      operationId: getSubscriberFeatures
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriberFeatureResponse"
    post:
      tags:
      - Management-data-controller
      operationId: createSubscriberFeature
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberFeatureUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubscriberFeatureResponse"
  /api/v1/additional-features:
    get:
      tags:
      - Management-data-controller
      operationId: getServices
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdditionalFeatureResponse"
    post:
      tags:
      - Management-data-controller
      operationId: createService
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdditionalFeatureUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdditionalFeatureResponse"
  /api/v1/feature-categories:
    get:
      tags:
      - Management-data-controller
      operationId: getServiceCategories
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IdNameResponse"
    post:
      tags:
      - Management-data-controller
      operationId: createServiceCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdNameResponse"
...
  /api/v1/subscribers/{subscriberId}/features:
    get:
      tags:
      - service-controller
      operationId: findSubscriberFeatures
      parameters:
      - name: subscriberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: query
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureSummaryDto"
  /api/v1/features/{featureId}:
    get:
      tags:
      - service-controller
      operationId: getFeatureDetails
      parameters:
      - name: featureId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeatureDetailsResponse"
...
components:
  schemas:
...
    NameRequest:
      type: object
      properties:
        name:
          type: string
      required:
      - name
    IdNameResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SubscriberUpsertRequest:
      type: object
      properties:
        phone:
          type: string
        fullName:
          type: string
        balance:
          type: number
          minimum: 0.00
        currentTariffId:
          type: integer
          format: int64
      required:
      - balance
      - fullName
      - phone
    SubscriberResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        phone:
          type: string
        fullName:
          type: string
        balance:
          type: number
        currentTariffId:
          type: integer
          format: int64
    SubscriberFeatureUpsertRequest:
      type: object
      properties:
        subscriberId:
          type: integer
          format: int64
        featureId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - ACTIVE
          - DISABLED
        connectedAt:
          type: string
          format: date-time
        disabledAt:
          type: string
          format: date-time
      required:
      - featureId
      - status
      - subscriberId
    SubscriberFeatureResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        subscriberId:
          type: integer
          format: int64
        featureId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - ACTIVE
          - DISABLED
        connectedAt:
          type: string
          format: date-time
        disabledAt:
          type: string
          format: date-time
    AdditionalFeatureUpsertRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        monthlyFee:
          type: number
          minimum: 0.00
        categoryId:
          type: integer
          format: int64
      required:
      - categoryId
      - description
      - monthlyFee
      - name
    AdditionalFeatureResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        monthlyFee:
          type: number
        categoryId:
          type: integer
          format: int64
...
    BillingTransactionDto:
      type: object
      properties:
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
...
    NotificationDto:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
        success:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    DisableFeatureResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        featureId:
          type: integer
          format: int64
        billingTransactions:
          type: array
          items:
            $ref: "#/components/schemas/BillingTransactionDto"
        notification:
          $ref: "#/components/schemas/NotificationDto"
...
    FeatureSummaryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          type: string
        monthlyFee:
          type: number
        status:
          type: string
    FeatureDetailsResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        category:
          type: string
        monthlyFee:
          type: number